pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-assistant-backend"
version = "0.1.0"
description = "Backend for AI Assistant application"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["ai", "assistant", "backend"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "langchain>=1.2.7",
    "langchain-ollama>=1.0.1",
    "langgraph-checkpoint-sqlite",
    "aiosqlite"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",  # For testing async endpoints
    "httpx>=0.24.0",  # For testing FastAPI with TestClient
    "black>=23.7.0",
    "ruff>=0.0.285",
    "mypy>=1.5.0",
    "pyinstaller>=6.0.0",  # For packaging as executable
]

otel = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",  # gRPC exporter
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",  # HTTP exporter
    "opentelemetry-instrumentation-fastapi>=0.41b0",   # Auto-instrument FastAPI
    "opentelemetry-instrumentation-logging>=0.41b0",   # Bridge Python logging to OTEL
]

[project.urls]
Homepage = "https://github.com/yourusername/ai-assistant-2"
Repository = "https://github.com/yourusername/ai-assistant-2"
Issues = "https://github.com/yourusername/ai-assistant-2/issues"

[tool.setuptools]
packages = {find = {where = ["."], include = ["src*"]}}
package-dir = {"" = "."}

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.mypy_cache
  | \.pytest_cache
  | __pycache__
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

